+++ src/components/wchar/wcsspn.c
@@ -20,5 +20,5 @@ size_t wcsspn(const wchar_t *s, const wchar_t *accept) {
 		++s_ptr;
 	}
 
-	return s_ptr - s;
+	return (size_t)(s_ptr - s);
 }
